# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * 833 Tidyups Booking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type BookingServiceType = typeof BookingServiceType[keyof typeof BookingServiceType];


export const BookingServiceType = {
  standard_clean: 'standard_clean',
  deep_clean: 'deep_clean',
  move_in_out: 'move_in_out',
  post_construction: 'post_construction',
} as const;

export type BookingFrequency = typeof BookingFrequency[keyof typeof BookingFrequency];


export const BookingFrequency = {
  one_time: 'one_time',
  weekly: 'weekly',
  biweekly: 'biweekly',
  monthly: 'monthly',
} as const;

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  firstName: string;
  lastName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  address: string;
  city: string;
  province?: string;
  /** @nullable */
  postalCode?: string | null;
  serviceType: BookingServiceType;
  bedrooms: number;
  bathrooms: number;
  extras?: string[];
  /** Date in YYYY-MM-DD format */
  scheduledDate: string;
  /** Time in HH:MM format */
  scheduledTime: string;
  frequency: BookingFrequency;
  /** @nullable */
  estimatedPrice?: number | null;
  /** @nullable */
  notes?: string | null;
  status: BookingStatus;
  /** @nullable */
  jobberJobId?: string | null;
  createdAt: string;
}

export type BookingInputServiceType = typeof BookingInputServiceType[keyof typeof BookingInputServiceType];


export const BookingInputServiceType = {
  standard_clean: 'standard_clean',
  deep_clean: 'deep_clean',
  move_in_out: 'move_in_out',
  post_construction: 'post_construction',
} as const;

export type BookingInputFrequency = typeof BookingInputFrequency[keyof typeof BookingInputFrequency];


export const BookingInputFrequency = {
  one_time: 'one_time',
  weekly: 'weekly',
  biweekly: 'biweekly',
  monthly: 'monthly',
} as const;

export type BookingInputStatus = typeof BookingInputStatus[keyof typeof BookingInputStatus];


export const BookingInputStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface BookingInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 7 */
  phone: string;
  email?: string;
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  city: string;
  province?: string;
  postalCode?: string;
  serviceType: BookingInputServiceType;
  /**
     * @minimum 0
     * @maximum 10
     */
  bedrooms: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  bathrooms: number;
  extras?: string[];
  scheduledDate: string;
  scheduledTime: string;
  frequency: BookingInputFrequency;
  estimatedPrice?: number;
  notes?: string;
  status?: BookingInputStatus;
}

export type BookingUpdateServiceType = typeof BookingUpdateServiceType[keyof typeof BookingUpdateServiceType];


export const BookingUpdateServiceType = {
  standard_clean: 'standard_clean',
  deep_clean: 'deep_clean',
  move_in_out: 'move_in_out',
  post_construction: 'post_construction',
} as const;

export type BookingUpdateFrequency = typeof BookingUpdateFrequency[keyof typeof BookingUpdateFrequency];


export const BookingUpdateFrequency = {
  one_time: 'one_time',
  weekly: 'weekly',
  biweekly: 'biweekly',
  monthly: 'monthly',
} as const;

export type BookingUpdateStatus = typeof BookingUpdateStatus[keyof typeof BookingUpdateStatus];


export const BookingUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface BookingUpdate {
  firstName?: string;
  lastName?: string;
  phone?: string;
  email?: string;
  address?: string;
  city?: string;
  province?: string;
  postalCode?: string;
  serviceType?: BookingUpdateServiceType;
  bedrooms?: number;
  bathrooms?: number;
  extras?: string[];
  scheduledDate?: string;
  scheduledTime?: string;
  frequency?: BookingUpdateFrequency;
  estimatedPrice?: number;
  notes?: string;
  status?: BookingUpdateStatus;
  jobberJobId?: string;
}

export interface BookingStats {
  totalBookings: number;
  upcomingCount: number;
  completedCount: number;
  pendingCount: number;
  cancelledCount: number;
  totalRevenue: number;
  thisWeekCount: number;
  thisMonthCount: number;
}

export type ListBookingsParams = {
status?: ListBookingsStatus;
limit?: number;
offset?: number;
};

export type ListBookingsStatus = typeof ListBookingsStatus[keyof typeof ListBookingsStatus];


export const ListBookingsStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;