# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * 833 Tidyups Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingFrequency } from './bookingFrequency';
import type { BookingServiceType } from './bookingServiceType';
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  firstName: string;
  lastName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  address: string;
  city: string;
  province?: string;
  /** @nullable */
  postalCode?: string | null;
  serviceType: BookingServiceType;
  bedrooms: number;
  bathrooms: number;
  extras?: string[];
  /** Date in YYYY-MM-DD format */
  scheduledDate: string;
  /** Time in HH:MM format */
  scheduledTime: string;
  frequency: BookingFrequency;
  /** @nullable */
  estimatedPrice?: number | null;
  /** @nullable */
  notes?: string | null;
  status: BookingStatus;
  /** @nullable */
  jobberJobId?: string | null;
  createdAt: Date;
}