# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * 833 Tidyups Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingInputFrequency } from './bookingInputFrequency';
import type { BookingInputServiceType } from './bookingInputServiceType';
import type { BookingInputStatus } from './bookingInputStatus';

export interface BookingInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 7 */
  phone: string;
  email?: string;
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  city: string;
  province?: string;
  postalCode?: string;
  serviceType: BookingInputServiceType;
  /**
     * @minimum 0
     * @maximum 10
     */
  bedrooms: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  bathrooms: number;
  extras?: string[];
  scheduledDate: string;
  scheduledTime: string;
  frequency: BookingInputFrequency;
  estimatedPrice?: number;
  notes?: string;
  status?: BookingInputStatus;
}