# lib/api-zod/src/generated/types/bookingStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * 833 Tidyups Booking API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingStats {
  totalBookings: number;
  upcomingCount: number;
  completedCount: number;
  pendingCount: number;
  cancelledCount: number;
  totalRevenue: number;
  thisWeekCount: number;
  thisMonthCount: number;
}