# lib/api-zod/src/generated/types/bookingUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * 833 Tidyups Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingUpdateFrequency } from './bookingUpdateFrequency';
import type { BookingUpdateServiceType } from './bookingUpdateServiceType';
import type { BookingUpdateStatus } from './bookingUpdateStatus';

export interface BookingUpdate {
  firstName?: string;
  lastName?: string;
  phone?: string;
  email?: string;
  address?: string;
  city?: string;
  province?: string;
  postalCode?: string;
  serviceType?: BookingUpdateServiceType;
  bedrooms?: number;
  bathrooms?: number;
  extras?: string[];
  scheduledDate?: string;
  scheduledTime?: string;
  frequency?: BookingUpdateFrequency;
  estimatedPrice?: number;
  notes?: string;
  status?: BookingUpdateStatus;
  jobberJobId?: string;
}